Move visitor to utility header
