+++ src/cpp/manipulator/override_manipulator.cpp
@@ -1,6 +1,6 @@
-
-#include <cse/manipulator.hpp>
-#include <cse/error.hpp>
+#include "cse/error.hpp"
+#include "cse/manipulator.hpp"
+#include "cse/utility/utility.hpp"
 
 #include <sstream>
 
@@ -9,17 +9,6 @@ namespace cse
 
 namespace
 {
-
-template<typename... Functors>
-struct visitor : Functors...
-{
-    visitor(const Functors&... functors)
-        : Functors(functors)...
-    {
-    }
-
-    using Functors::operator()...;
-};
 
 cse::variable_causality find_variable_causality(
     const std::vector<variable_description>& variables,
