+++ src/cpp/manipulator/scenario_manager.cpp
@@ -4,25 +4,10 @@
 #include "cse/log/logger.hpp"
 #include "cse/scenario.hpp"
 #include "cse/scenario_parser.hpp"
+#include "cse/utility/utility.hpp"
 
 namespace cse
 {
-
-namespace
-{
-
-template<typename... Functors>
-struct visitor : Functors...
-{
-    visitor(const Functors&... functors)
-        : Functors(functors)...
-    {
-    }
-
-    using Functors::operator()...;
-};
-
-} // namespace
 
 class scenario_manager::impl
 {
